+++ src/rec/audio/source/DoubleRunnyBuffer.h
@@ -26,7 +26,7 @@ class DoubleRunnyBuffer
 
   virtual PositionableAudioSource* makeSource();
   gui::CachedThumbnail* cachedThumbnail() { return cachedThumbnail_.get(); }
-  virtual int getTotalLength() const { return buffery_->getLength(); }
+  virtual int64 getTotalLength() const { return buffery_->getLength(); }
 
   virtual void operator()(const StretchyProto& p) { setStretchy(p); }
