+++ src/rec/audio/source/Runny.h
@@ -18,7 +18,7 @@ class Runny : public Thread, public Wrappy::Position {
 
   virtual ~Runny();
 
-  virtual void setNextReadPosition(int p);
+  virtual void setNextReadPosition(int64 p);
   virtual void getNextAudioBlock(const AudioSourceChannelInfo& info);
 
   // Try to pre-fill the lookahead buffer one slot.
